+++ tests/test_cli.py
@@ -116,6 +116,11 @@ def test_cli_default_detector():
     assert invoke_scenedetect("-i {VIDEO} time {TIME}", config_file=None) == 0
 
 
+def test_cli_crop():
+    """Test --crop functionality."""
+    assert invoke_scenedetect("-i {VIDEO} --crop 0 0 256 256 time {TIME}", config_file=None) == 0
+
+
 @pytest.mark.parametrize("info_command", ["help", "about", "version"])
 def test_cli_info_command(info_command):
     """Test `scenedetect` info commands (e.g. help, about)."""